Added isLoud function to check loudness of uploads
Wrote isLoud function that takes save_path argument and checks the loudness level using ffmpeg in a sub process call and compares it to a static value set inside the function. Returns True or False. Added isLoud to the add function and deletes if isLoud returns true, if not adds the sound.